moving out postprocess_logical_form to semantic_parsing_util file
